3rd Person Isometric Part 1
Edited the lasso behavior script to make the lasso's starting position and rotation public variables. Changed my scene to have a 3rd person isometric camera that tracks the player, and movement to match. Scripts exist for attack implementation but are unfinished. Currently, the player is knocked off course when colliding with an enemy, beginning to move without input, so their collider is a trigger for now.