Jukebox Rework and Barrel Sound
Reworked the Jukebox system so that now it is fed Audio Sources rather than just clips, giving us control over 3D sound properties and the like. This currently does not work when playing death/destruction sounds for objects that are then destroyed, but Josh has assured me that this problem will sort itself out as animations, etc. are added. The barrel explosion sound is also now in the project as part of testing.