12/4 balance changes. Check description or #progress-report in Design for full notes
Changelog 12/4

- Tombstone_throwable 3 (the wooden cross) health changed from 30 to 20
A merge conflict reverted this change from before. This simply restores the health back to what was intended previously

- Melee enemy Fire Damage per tick changed from 3 to 4
Another merge conflict fix

- Melee enemy dash attack damage increased from 8 to 10
Simple buff to increase lethality of the melee enemy

- Melee enemy Break out Time decreased from 5 seconds to 3 seconds
The melee enemy could stay stuck in the tendril for a bit too long. Reducing the time before it breaks out from 5 to 3 still allows the player to spend time making a decision with the grabbed enemy while having the decision itself act out faster

- Melee enemy attack frames per second decreased from 10 to 4
- Melee enemy dash range reduced from 10 to 8
- Melee dash time reduced from .75 seconds to .5 seconds
All of these changes were made to make the dash itself faster, and more interesting to fight against. In exchange for a faster dash, the range it can travel with it was slightly reduced from 10 to 8

- Melee enemy attack wind down seconds increased from 1 to 1.25
This gives a slightly bigger opening for the player to grab the melee enemy. It simulates the melee enemy swinging the pickaxe into the ground, missing, and strugging to pluck it out of the hard ground.

- Crate is no longer frozen in the X, Y, or Z rotation
- Crate Time to Stop increased from 1 to 1.5
These changes allow the crate to freely rotate and tumble when thrown, which makes the player feel stronger when throwing them, as opposed to simply sliding the crate across the floor

- Explosive barrel explosion damage increased from 35 to 50
Damage increased to instantly kill any basic melee or ranged enemy. This is to remove those feel bad moments when an explosion won't kill one of those enemies

- Added Playerdmg variable to the Explosive Barrel. The added code for this (In the ExplosionBehavior script) is below to double check the functionality. The variable is currently set to 20 since we want the player to take reduced damage from the explosion.

 if (damaged != null)
            {
                if (entity.gameObject.CompareTag("Player"))
                {
                    damaged.TakeDamage(Playerdmg);
                }

                else
                {
                    damaged.TakeDamage(dmg);
                }