Dialog UI Updates
Moved dialogue options to be part of the larger text box and turned the button opacity to zero, switching the color changing functionality to affect the text itself rather than the button image. Adjusted sizing and positioning of NPC portrait and the image feature. Made the dialogue options a vertical layout group