Finishing Up Dialogue UI Implementation
Everything works, basically! Only thing that doesn't is that the nametag colors for NPCs change correctly but the splashes behind them do not. Town collectibles should mostly not be affected, but a few might need to be tweaked over to the temp note UI and/or have their Ink files adjusted.